# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/cards": {
            "get": {
                "description": "Get  a list of cards",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "/cards"
                ],
                "summary": "Get  a list of cards",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controllers.CardsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controllers.HTTPError"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new card item",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "/cards"
                ],
                "summary": "Create a card",
                "parameters": [
                    {
                        "description": "Add a New Card Request",
                        "name": "CardRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.CardRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controllers.CardResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controllers.HTTPError"
                        }
                    }
                }
            }
        },
        "/cards/{id}": {
            "get": {
                "description": "Get details of card corresponding to the input cardId",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "/cards"
                ],
                "summary": "Get details for a given cardId",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID of the card",
                        "name": "cardId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controllers.CardResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controllers.HTTPError"
                        }
                    }
                }
            },
            "put": {
                "description": "Update the card corresponding to the input cardId",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "/cards"
                ],
                "summary": "Update card identified by the given cardId",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID of the card to be updated",
                        "name": "cardId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controllers.CardResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controllers.HTTPError"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a new card item",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "/cards"
                ],
                "summary": "Delete a card",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID of the card to be deleted",
                        "name": "cardId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/controllers.HTTPError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "controllers.CardRequest": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "Cards for this user",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.ReqAddCard"
                        }
                    ]
                }
            }
        },
        "controllers.CardResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "Cards for this user",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.Card"
                        }
                    ]
                },
                "message": {
                    "description": "Message of the response\nin: string",
                    "type": "string"
                },
                "status": {
                    "description": "Status of the error\nin: int64",
                    "type": "integer"
                }
            }
        },
        "controllers.CardsResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Card"
                    }
                },
                "message": {
                    "description": "Message of the response\nin: string",
                    "type": "string"
                },
                "status": {
                    "description": "Status of the error\nin: int64",
                    "type": "integer"
                }
            }
        },
        "controllers.HTTPError": {
            "type": "object",
            "properties": {
                "message": {
                    "description": "Message of the error\nin: string",
                    "type": "string"
                },
                "status": {
                    "description": "Status of the error\nin: int64",
                    "type": "integer"
                }
            }
        },
        "models.Card": {
            "type": "object",
            "properties": {
                "description": {
                    "description": "Description of the card\nin: string",
                    "type": "string"
                },
                "id": {
                    "description": "Id of the card\nin: int64",
                    "type": "integer"
                },
                "name": {
                    "description": "Name of the card\nin: string",
                    "type": "string"
                },
                "status": {
                    "description": "Status of the card\nin: int64",
                    "type": "integer"
                }
            }
        },
        "models.ReqAddCard": {
            "type": "object",
            "required": [
                "description",
                "name"
            ],
            "properties": {
                "description": {
                    "description": "Description of the card\nin: string",
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 2
                },
                "name": {
                    "description": "Name of the card\nin: string",
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 2
                }
            }
        }
    }
}